Move open to launcher module
