+++ jdaviz/core/data_formats.py
@@ -11,8 +11,6 @@ from specutils import Spectrum1D, SpectrumList, SpectrumCollection
 from stdatamodels import asdf_in_fits
 
 from jdaviz.core.config import list_configurations
-from jdaviz import configs as jdaviz_configs
-from jdaviz.cli import DEFAULT_VERBOSITY, DEFAULT_HISTORY_VERBOSITY
 
 __all__ = [
     'guess_dimensionality',
@@ -279,67 +277,3 @@ def identify_helper(filename, ext=1):
         return (['imviz'], hdul)
 
     raise ValueError(f"No helper could be auto-identified for {filename}.")
-
-
-def open(filename, show=True, **kwargs):
-    '''
-    Automatically detect the correct configuration based on a given file,
-    load the data, and display the configuration
-
-    Parameters
-    ----------
-    filename : str (path-like)
-        Name for a local data file.
-    show : bool
-        Determines whether to immediately show the application
-
-    All other arguments are interpreted as load_data arguments for
-    the autoidentified configuration class
-
-    Returns
-    -------
-    Jdaviz ConfigHelper : jdaviz.core.helpers.ConfigHelper
-        The autoidentified ConfigHelper for the given data
-    '''
-    # Identify the correct config
-    helper_str, hdul = identify_helper(filename)
-    return _launch_config_with_data(helper_str, hdul, show, **kwargs)
-
-
-def _launch_config_with_data(config, data=None, show=True, **kwargs):
-    '''
-    Launch jdaviz with a specific, known configuration and data
-
-    Parameters
-    ----------
-    config : str (path-like)
-        Name for a local data file.
-    data : str or any Jdaviz-compatible data
-        A filepath or Jdaviz-compatible data object (such as Spectrum1D or CCDData)
-    show : bool
-        Determines whether to immediately show the application
-
-    All other arguments are interpreted as load_data/load_spectrum arguments for
-    the autoidentified configuration class
-
-    Returns
-    -------
-    Jdaviz ConfigHelper : jdaviz.core.helpers.ConfigHelper
-        The loaded ConfigHelper with data loaded
-    '''
-    viz_class = getattr(jdaviz_configs, config.capitalize())
-
-    # Create config instance
-    verbosity = kwargs.pop('verbosity', DEFAULT_VERBOSITY)
-    history_verbosity = kwargs.pop('history_verbosity', DEFAULT_HISTORY_VERBOSITY)
-    viz_helper = viz_class(verbosity=verbosity, history_verbosity=history_verbosity)
-
-    # Load data
-    if data not in (None, ''):
-        viz_helper.load_data(data, **kwargs)
-
-    # Display app
-    if show:
-        viz_helper.show()
-
-    return viz_helper
