+++ jdaviz/core/launcher.py
@@ -1,8 +1,77 @@
 import ipyvuetify as v
 from ipywidgets import jslink
 
-from jdaviz.cli import ALL_JDAVIZ_CONFIGS
-from jdaviz.core.data_formats import identify_helper, _launch_config_with_data
+from jdaviz import configs as jdaviz_configs
+from jdaviz.cli import DEFAULT_VERBOSITY, DEFAULT_HISTORY_VERBOSITY, ALL_JDAVIZ_CONFIGS
+from jdaviz.core.data_formats import identify_helper
+
+
+def open(filename, show=True, **kwargs):
+    '''
+    Automatically detect the correct configuration based on a given file,
+    load the data, and display the configuration
+
+    Parameters
+    ----------
+    filename : str (path-like)
+        Name for a local data file.
+    show : bool
+        Determines whether to immediately show the application
+
+    All other arguments are interpreted as load_data/load_spectrum arguments for
+    the autoidentified configuration class
+
+    Returns
+    -------
+    Jdaviz ConfigHelper : jdaviz.core.helpers.ConfigHelper
+        The autoidentified ConfigHelper for the given data
+    '''
+    # Identify the correct config
+    compatible_helpers, hdul = identify_helper(filename)
+    if len(compatible_helpers) > 1:
+        raise NotImplementedError(f"Multiple helpers provided: {compatible_helpers}."
+                                  "Unsure which to launch")
+    else:
+        return _launch_config_with_data(compatible_helpers[0], hdul, show, **kwargs)
+
+
+def _launch_config_with_data(config, data=None, show=True, **kwargs):
+    '''
+    Launch jdaviz with a specific, known configuration and data
+
+    Parameters
+    ----------
+    config : str (path-like)
+        Name for a local data file.
+    data : str or any Jdaviz-compatible data
+        A filepath or Jdaviz-compatible data object (such as Spectrum1D or CCDData)
+    show : bool
+        Determines whether to immediately show the application
+
+    All other arguments are interpreted as load_data/load_spectrum arguments for
+    the autoidentified configuration class
+
+    Returns
+    -------
+    Jdaviz ConfigHelper : jdaviz.core.helpers.ConfigHelper
+        The loaded ConfigHelper with data loaded
+    '''
+    viz_class = getattr(jdaviz_configs, config.capitalize())
+
+    # Create config instance
+    verbosity = kwargs.pop('verbosity', DEFAULT_VERBOSITY)
+    history_verbosity = kwargs.pop('history_verbosity', DEFAULT_HISTORY_VERBOSITY)
+    viz_helper = viz_class(verbosity=verbosity, history_verbosity=history_verbosity)
+
+    # Load data
+    if data not in (None, ''):
+        viz_helper.load_data(data, **kwargs)
+
+    # Display app
+    if show:
+        viz_helper.show()
+
+    return viz_helper
 
 
 def show_launcher(configs=['imviz', 'specviz', 'mosviz', 'cubeviz', 'specviz2d']):
